use jsdom-worker instead of mock
