+++ client/babel.config.json
@@ -2,7 +2,7 @@
     "presets": ["@babel/preset-env", "@babel/preset-typescript"],
     "env": {
         "test": {
-            "plugins": ["@babel/plugin-transform-runtime"]
+            "plugins": ["@babel/plugin-transform-runtime", "babel-plugin-transform-import-meta"]
         }
     }
 }
