+++ client/package.json
@@ -154,6 +154,7 @@
     "autoprefixer": "10.4.13",
     "axios-mock-adapter": "^1.20.0",
     "babel-jest": "^29.3.1",
+    "babel-plugin-transform-import-meta": "^2.2.0",
     "buffer": "^6.0.3",
     "cpy-cli": "^4.2.0",
     "css-loader": "^6.7.3",
@@ -170,6 +171,7 @@
     "jest": "^29.3.1",
     "jest-environment-jsdom": "^29.3.1",
     "jest-location-mock": "^1.0.9",
+    "jsdom-worker": "^0.3.0",
     "json-loader": "^0.5.7",
     "mini-css-extract-plugin": "^2.5.3",
     "postcss-loader": "^7.0.2",
