--- client/src/composables/filter/__mocks__/index.ts
@@ -1,1 +1,0 @@
-export { useFilterObjectArray } from "./filter";
